Détection de maladie/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.16.2
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
huggingface_hub==0.30.2
